# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Translations",
        "version": "1.0.0"
    },
    "paths": {
        "/api/translations": {
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Store a new translation",
                "description": "Store a new translation for a given meta_key and key.",
                "operationId": "fe38d8cf02ff5c92d2cd8d5dfb095e5f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "language_id",
                                    "meta_key": "greetings",
                                    "content": "[1,2,3]"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "language_id": {
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "meta_key": {
                                                    "type": "string",
                                                    "example": "greetings"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "text"
                                                },
                                                "tags": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "number",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "name"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "message": {
                                                    "properties": {
                                                        "language_id": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string",
                                                                "example": "The language id field is required."
                                                            },
                                                            "collectionFormat": "multi"
                                                        },
                                                        "meta_key": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string",
                                                                "example": "The meta key field is required."
                                                            },
                                                            "collectionFormat": "multi"
                                                        },
                                                        "content": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string",
                                                                "example": "The contentfield is required."
                                                            },
                                                            "collectionFormat": "multi"
                                                        },
                                                        "tags": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string",
                                                                "example": "The tags field must be an array."
                                                            },
                                                            "collectionFormat": "multi"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/translations-export": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Get all translations",
                "description": "Get all translations",
                "operationId": "524a957db3e6b2a6b26f7efe2520c9b1",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "if page parameter is not added it will be default 1",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "language_id": {
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "meta_key": {
                                                    "type": "string",
                                                    "example": "greetings"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "text"
                                                },
                                                "tags": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "number",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "name"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/translations-search": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Search translation by meta_key or tag",
                "description": "Search translation by meta_key or tag",
                "operationId": "b861d1e84c30943e4f05992af17a6416",
                "parameters": [
                    {
                        "name": "meta_key",
                        "in": "path",
                        "description": "If the search query is based on a <strong>meta_key</strong>, the search will be filtered by the <strong>meta_key</strong>.",
                        "schema": {
                            "type": "string"
                        },
                        "example": "greetings"
                    },
                    {
                        "name": "tag",
                        "in": "path",
                        "description": "If the search query is based on a <strong>tag</strong>, the search will be filtered by the <strong>tag</strong>.",
                        "schema": {
                            "type": "string"
                        },
                        "example": "mobile"
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "description": "If page parameter is not added it will be default 1",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "language_id": {
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "meta_key": {
                                                    "type": "string",
                                                    "example": "greetings"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "text"
                                                },
                                                "tags": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "number",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "name"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}